Review221005

+++ Models/Public/Class6_7_C.cs
@@ -8,6 +8,6 @@ namespace GoWMS.Server.Models.Public
     public class Class6_7_C
     {
         public DateTime? W_Hour { get; set; }
-        public Int64? W_Count { get; set; }
+        public Int32? W_Count { get; set; }
     }
 }
