+++ Models/Public/Class6_7_F.cs
@@ -11,14 +11,14 @@ namespace GoWMS.Server.Models.Public
         [DataType(DataType.Date)]
         [DisplayFormat(ApplyFormatInEditMode = true, DataFormatString = "{0:dd/MM/yyyy}")]
         public DateTime? W_date { get; set; }
-        public Int64? W01 { get; set; }
-        public Int64? W07 { get; set; }
-        public Int64? W02 { get; set; }
-        public Int64? W03 { get; set; }
-        public Int64? W05 { get; set; }
-        public Int64? W09 { get; set; }
-        public Int64? W101 { get; set; }
-        public Int64? W102 { get; set; }
-        public Int64? Wtotal { get; set; }
+        public Int32? W01 { get; set; }
+        public Int32? W07 { get; set; }
+        public Int32? W02 { get; set; }
+        public Int32? W03 { get; set; }
+        public Int32? W05 { get; set; }
+        public Int32? W09 { get; set; }
+        public Int32? W101 { get; set; }
+        public Int32? W102 { get; set; }
+        public Int32? Wtotal { get; set; }
     }
 }
